+++ apps/angular-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # angular-example
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/devtools@4.0.2
+  - @tryabby/angular@1.0.1
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/devtools@4.0.1
+  - @tryabby/angular@1.0.1
+
 ## 0.0.5
 
 ### Patch Changes